Added help string message function Added function to return map name from map id Fixed colorLerp not returning correct data type

+++ include/color.h
@@ -41,7 +41,7 @@
  * 
  * AUTHOR :			Daniel "Dnawrkshp" Gerendasy
  */
-u64 colorLerp(u32 a, u32 b, float t);
+u32 colorLerp(u32 a, u32 b, float t);
 
 #endif // _LIBDL_COLOR_H_
